11/21 Changes
When max short wall distance reached, will go home.
All values for the cases we tested have been optimized to the best of my
knowledge.
We still need to account for the case where the object is just to the
right of the bot as it sweeps. I did some testing and the bot just rams
into the object and stops. I think we should implement a separate
routine that sees if the dist5 is less than a certain value and has is
just turn, tap and move backwards.